Create Check Anagram or not
Using Hash Map or Dictionary
The idea is to use a hash map or dictionary count the frequency of each character in both the input strings. If the frequency of every character matches in both strings, then the strings are anagrams.

First, count the occurrences of each character in first string.
Then, decrement the count for each character in the second string.
If the strings are anagrams, all positions in the frequency array should be zero as any non-zero position means that the frequency of that character is not same in both strings.