Create Aggresive cows
Given an array stalls[] which denotes the position of a stall and an integer k which denotes the number of aggressive cows.

 The task is to assign stalls to k cows such that the minimum distance between any two of them is the maximum possible.

Using Binary Search
The minimum distance between the cows has a monotonic property.

If we can place all the cows with a minimum distance d, then we can also place them with any smaller distance than d. The reason is by reducing the minimum gap, we can place more number of cows as lower gaps allow us to place cows even more closer.
If we can't place all the cows with a minimum distance d, then we can't place them with any larger distance than d. The reason is if the gap is already too large to place the cows, then larger gap will also not work.
Therefore, we can use binary search to find the maximum possible minimum distance, and to check the for any distance, we place the first cow in the first stall and the next ones only if the gap from the last placed cow is at least that distance. If all cows can be placed for a certain distance, then it is a feasible distance.