Create  two Sum-pair with given sum
Using Hash Set - O(n) time and O(n) space
Hashing provides a more efficient solution to the 2Sum problem. Rather than checking every possible pair, we store each number in an unordered set during iterating over the array's elements. For each number, we calculate its complement (i.e., target - current number) and check if this complement exists in the set. If it does, we have successfully found the pair that sums to the target. This approach significantly reduces the time complexity and allowing us to solve the problem in linear time O(n).

Step-by-step approach:

Create an empty Hash Set or Unordered Set
Iterate through the array and for each number in the array:
Calculate the complement (target - current number).
Check if the complement exists in the set:
If it is, then pair found.
If it isn’t, add the current number to the set.
If the loop completes without finding a pair, return that no pair exists.