Create Count Pairs whose sum is strictly less than target
Using Two Pointers Technique - O(n*logn+n) Time and O(1) Space

First sort the array, then use Two Pointers Technique to find the number of pairs with a sum less than the given target. 

Initialize two pointers, one at the beginning (left) and the other at the end (right) of the array. Now, compare the sum of elements at these pointers with the target:

If sum < target:
Pairs formed by the element at the left pointer with every element between left and right (inclusive) will have a sum less than the target. 
Therefore, we add (right - left) to the count and move the left pointer one step to the right to explore more pairs.
If sum >= target:
We move the right pointer one step to the left to reduce the sum