Create LRU Cache Implementation using Doubly Linked List
The idea is to keep inserting the key-value pair at the head of the doubly linked list. When a node is accessed or added, it is moved to the head of the list (right after the dummy head node). This marks it as the most recently used. When the cache exceeds its capacity, the node at the tail (right before the dummy tail node) is removed as it represents the least recently used item.

Doubly-Linked-List-Representation-Of-LRU-Cache-5.webp