Create Valid Parentheses in an Expression
he idea is to put all the opening brackets in the stack. Whenever you hit a closing bracket, search if the top of the stack is the opening bracket of the same nature. If this holds then pop the stack and continue the iteration. In the end if the stack is empty, it means all brackets are balanced or well-formed. Otherwise, they are not balanced.

Step-by-step approach:

Declare a character stack (say temp).
Now traverse the string s. 
If the current character is an opening bracket ( '(' or '{'  or '[' ) then push it to stack.
If the current character is a closing bracket ( ')' or '}' or ']' ) and the closing bracket matches with the opening bracket at the top of stack, then pop the opening bracket. Else s is not balanced.
After complete traversal, if some starting brackets are left in the stack then the expression is not balanced, else balanced.